casynth: panic: Reset node dead state
Note activation depends on their note_dead state. Only "dead" note is
playable, otherwise it will be considered as a stale note, and to be
cleaned up. A common NOTE_OFF event will reset this state.

If not reset on panic as well, the notes which are activated on a
ALL_NOTES_OFF call will unable to be played again.

For example, acquire ALL_NOTES_OFF when playing note C4 in sustain,
then if you play C4 again, it goes mute while other notes are still
available.

So it is necessary to reset note_dead state.