casynth: FLTK: Rename show() into init()
Do not mix UI initializer and show method into one function. Split them.
This clarifies the logic of program, letting the functions to do what
they should do, respectively.

This is necessary for implementing embed window with FLTK's own native
implementations (use Fl_Window::parent() to set parent widget). Parent
window needed to be set before the window is shown. In this situation,
mixing "init" and "show" in show() method is not good for me.

See:
- Fl_WinAPI_Window_Driver::makeWindow()
- Fl_Widget::parent(Fl_Group* p)