casynth: Gracefully handles the exit of plugin UI.
On DPF, it's up to developers to handle the process of closing external
window.

In many 3rd-party toolkits (Qt, FLTK, NTK, etc.), simply closing the
window does not mean to close the DISTRHO::UI instance. Instead, the
instance is still active on background, even though the host is already
closed.

To entirely close the plugin instance, we need to call hide() in
DISTRHO::UI instance. This asks DPF to close and clean up explicitly.

In CaSynth, I registered a callback on main window close. The callback
will automatically invoke DISTRHO::UI::hide(), so if I close the UI in
DAW, or standalone JACK application, the plugin can close and clean up
properly.